Cloud_Functions_PubSub/process-order-placed/requirements.txt
google-cloud-pubsub
google-cloud-firestore
google-cloud-logging
google-cloud-bigquery==3.14.0